+++ patches/src/main/kotlin/app/morphe/patches/music/ad/video/HideVideoAds.kt
@@ -24,6 +24,7 @@ val hideVideoAdsPatch = bytecodePatch(
             "7.29.52",
             "8.10.52",
             "8.37.56",
+            "8.40.54",
         )
     )