+++ patches/src/main/kotlin/app/morphe/patches/music/misc/backgroundplayback/BackgroundPlaybackPatch.kt
@@ -19,6 +19,7 @@ val backgroundPlaybackPatch = bytecodePatch(
             "7.29.52",
             "8.10.52",
             "8.37.56",
+            "8.40.54",
         )
     )