+++ patches/src/main/kotlin/app/morphe/patches/music/misc/gms/GmsCoreSupportPatch.kt
@@ -31,6 +31,7 @@ val gmsCoreSupportPatch = gmsCoreSupportPatch(
             "7.29.52",
             "8.10.52",
             "8.37.56",
+            "8.40.54",
         )
     )
 }