+++ patches/src/main/kotlin/app/morphe/patches/youtube/ad/video/VideoAdsPatch.kt
@@ -32,6 +32,7 @@ val videoAdsPatch = bytecodePatch(
             "20.26.46",
             "20.31.42",
             "20.37.48",
+            "20.40.45",
         )
     )