+++ patches/src/main/kotlin/app/morphe/patches/youtube/misc/playservice/VersionCheckPatch.kt
@@ -101,7 +101,11 @@ var is_20_34_or_greater : Boolean by Delegates.notNull()
     private set
 var is_20_37_or_greater : Boolean by Delegates.notNull()
     private set
+var is_20_38_or_greater : Boolean by Delegates.notNull()
+    private set
 var is_20_39_or_greater : Boolean by Delegates.notNull()
+    private set
+var is_20_40_or_greater : Boolean by Delegates.notNull()
     private set
 var is_20_41_or_greater : Boolean by Delegates.notNull()
     private set
@@ -163,7 +167,9 @@ val versionCheckPatch = resourcePatch(
         is_20_31_or_greater = 253205000 <= playStoreServicesVersion
         is_20_34_or_greater = 253505000 <= playStoreServicesVersion
         is_20_37_or_greater = 253805000 <= playStoreServicesVersion
+        is_20_38_or_greater = 253905000 <= playStoreServicesVersion
         is_20_39_or_greater = 253980000 <= playStoreServicesVersion
+        is_20_40_or_greater = 254105000 <= playStoreServicesVersion
         is_20_41_or_greater = 254205000 <= playStoreServicesVersion
         is_20_43_or_greater = 254405000 <= playStoreServicesVersion
         is_20_45_or_greater = 254605000 <= playStoreServicesVersion